feat(my-egg-list): 심은 알 모달 위치 정보 표시 추가 및 발견 시간 아이콘 시계로 변경
